задание 1, раздел 5
